Move handler to be under `ee/idFed`
